% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCat.R
\docType{class}
\name{Cat-class}
\alias{Cat-class}
\title{class S4 Cat}
\description{
class S4 Cat
}
\section{Slots}{

\describe{
\item{\code{id}}{character. ID of the cat, composed of first letters of coat_pattern, coat_color and eyes_color
and the datetime at the moment of creation}

\item{\code{size}}{character. Size of the cat, between "kitten", "small", "medium", "big"}

\item{\code{fluffy}}{logical. Is the cat fluffy ?}

\item{\code{coat_pattern}}{character. Pattern of the coat}

\item{\code{coat_color}}{character. Color of the coat}

\item{\code{eyes_color}}{character. Eyes color of the cat}

\item{\code{behavior}}{character. Behavior type of the cat}

\item{\code{comments}}{character. Any comments of the user}

\item{\code{location}}{data.frame. date, lat and longituude of each occurence of the cat}
}}

