% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_requests.R
\name{create_table_locations}
\alias{create_table_locations}
\title{Create table Locations}
\usage{
create_table_locations(db)
}
\arguments{
\item{db}{database connexion}
}
\description{
Create table Locations
}
