% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_requests.R
\name{exists_email}
\alias{exists_email}
\title{Select user}
\usage{
exists_email(db, email)
}
\arguments{
\item{db}{database connexion}

\item{email}{character. ID of the cat}
}
\value{
data.frame
}
\description{
Select user
}
