% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cats.R
\name{get_cats}
\alias{get_cats}
\title{Title}
\usage{
get_cats(db, user_id)
}
\arguments{
\item{db}{database connexion}

\item{user_id}{character. ID of the user}
}
\value{
list of cats
}
\description{
Title
}
