% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_locations.R
\name{get_locations}
\alias{get_locations}
\title{Get Locations}
\usage{
get_locations(db, cat_id)
}
\arguments{
\item{db}{database connexion}

\item{cat_id}{character. ID of the cat}
}
\value{
list of location
}
\description{
Get Locations
}
