% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_requests.R
\name{insert_into_cats}
\alias{insert_into_cats}
\title{Insert cat into database}
\usage{
insert_into_cats(
  db,
  id,
  size,
  fluffy,
  coat_pattern,
  coat_color,
  eyes_color,
  behavior,
  comments
)
}
\arguments{
\item{db}{database connexion}

\item{id}{character. ID of the cat}

\item{size}{character. Size of the cat, between "kitten", "small", "medium", "big"}

\item{fluffy}{logical. Is the cat fluffy ?}

\item{coat_pattern}{character. Pattern of the coat}

\item{coat_color}{character. Color of the coat}

\item{eyes_color}{character. Eyes color of the cat}

\item{behavior}{character. Behavior type of the cat}

\item{comments}{character. Any comments of the user}
}
\description{
Insert cat into database
}
