% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_requests.R
\name{insert_into_location}
\alias{insert_into_location}
\title{Insert Into Location}
\usage{
insert_into_location(db, cat_id, seen_at, lat, lng)
}
\arguments{
\item{db}{database connexion}

\item{cat_id}{character. ID of the cat}

\item{seen_at}{time. date time when the cas is seen}

\item{lat}{numeric. lattitude of the cat}

\item{lng}{numeric. longitude of the cat}
}
\description{
Insert Into Location
}
