% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_requests.R
\name{insert_into_users}
\alias{insert_into_users}
\title{Insert Into Users}
\usage{
insert_into_users(db, user_id, name, email, password)
}
\arguments{
\item{db}{database connexion}

\item{user_id}{character. ID of the user, generated by uuid}

\item{name}{character. name chosen by the user}

\item{email}{character. email of the user, must not already exist in database}

\item{password}{character. password chosen by the user}
}
\description{
Insert Into Users
}
