% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classLocation.R
\docType{class}
\name{Location-class}
\alias{Location-class}
\title{Location}
\description{
Location
}
\section{Slots}{

\describe{
\item{\code{seen_at}}{Date. Date time of the location}

\item{\code{lat}}{numeric. lattitude of the location}

\item{\code{lng}}{numeric. longitude of the location}
}}

