% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_in.R
\name{log_in}
\alias{log_in}
\title{Log In}
\usage{
log_in(db, email, password)
}
\arguments{
\item{db}{database connexion}

\item{email}{character. email of the user, must not already exist in database}

\item{password}{character. password chosen by the user}
}
\value{
User object
}
\description{
Log In
}
