% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_cat.R
\name{new_cat}
\alias{new_cat}
\title{Create new Cat}
\usage{
new_cat(
  db,
  user_id,
  body_size,
  fluffy,
  coat_pattern,
  coat_color,
  eyes_color,
  behavior,
  comments,
  seen_at,
  lat,
  lng
)
}
\arguments{
\item{db}{database connexion}

\item{user_id}{character. ID of the user, generated by uuid}

\item{body_size}{character. body_size of the cat, between "kitten", "small", "medium", "big"}

\item{fluffy}{logical. Is the cat fluffy ?}

\item{coat_pattern}{character. Pattern of the coat}

\item{coat_color}{character. Color of the coat}

\item{eyes_color}{character. Eyes color of the cat}

\item{behavior}{character. Behavior type of the cat}

\item{comments}{character. Any comments of the user}

\item{seen_at}{time. date time when the cas is seen}

\item{lat}{numeric. lat of the cat}

\item{lng}{numeric. lng of the cat}
}
\value{
Cat
}
\description{
Create new Cat
}
