% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_location.R
\name{new_location}
\alias{new_location}
\title{New location}
\usage{
new_location(db, cat_id, seen_at, lat, lng)
}
\arguments{
\item{db}{database connexion}

\item{cat_id}{character. ID of the cat}

\item{seen_at}{time. date time when the cas is seen}

\item{lat}{numeric. lattitude of the cat}

\item{lng}{numeric. longitude of the cat}
}
\value{
Location object
}
\description{
New location
}
