% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_user.R
\name{new_user}
\alias{new_user}
\title{new User}
\usage{
new_user(db, name, email, password)
}
\arguments{
\item{db}{database connexion}

\item{name}{character. name chosen by the user}

\item{email}{character. email of the user, must not already exist in database}

\item{password}{character. password chosen by the user}
}
\value{
User object
}
\description{
new User
}
