% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_catname.R
\name{random_catname}
\alias{random_catname}
\title{Random Cat Name}
\usage{
random_catname(language = "EN")
}
\arguments{
\item{language}{two strings character for cats' language name}
}
\value{
character size 1
}
\description{
Random Cat Name
}
