% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_requests.R
\name{select_cats}
\alias{select_cats}
\title{Select cats}
\usage{
select_cats(db, user_id)
}
\arguments{
\item{db}{database connexion}

\item{user_id}{character. ID of the user}
}
\value{
data.frame
}
\description{
Select cats
}
