% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_requests.R
\name{select_locations}
\alias{select_locations}
\title{Select locations}
\usage{
select_locations(db, cat_id)
}
\arguments{
\item{db}{database connexion}

\item{cat_id}{character. ID of the cat}
}
\value{
data.frame
}
\description{
Select locations
}
