% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_requests.R
\name{select_user}
\alias{select_user}
\title{Select user}
\usage{
select_user(db, email)
}
\arguments{
\item{db}{database connexion}

\item{email}{character. ID of the cat}
}
\value{
data.frame
}
\description{
Select user
}
