% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classUser.R
\docType{class}
\name{User-class}
\alias{User-class}
\title{User}
\description{
User
}
\section{Slots}{

\describe{
\item{\code{seen_at}}{Date. Date time of the location}

\item{\code{lat}}{numeric. lattitude of the location}

\item{\code{lng}}{numeric. longitude of the location}
}}

