% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classUser.R
\docType{class}
\name{User-class}
\alias{User-class}
\title{User}
\description{
User
}
\section{Slots}{

\describe{
\item{\code{user_id}}{character. ID of the user, generated by uuid}

\item{\code{name}}{character. Name chosen by the user}

\item{\code{email}}{character. Email of the user, must be unique}

\item{\code{password}}{character. crypted password with bcrypt}

\item{\code{cats}}{list. List of cats of the user}
}}

